test: add unit test for new option
